+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 
 #### :rocket: New Feature
 
+- Add `littleEndian` feature for `DataView` to Stdlib. https://github.com/rescript-lang/rescript/pull/7881
+
 #### :bug: Bug fix
 
 - Include `-runtime-path` as bsc argument when generating `.mlmap` file. https://github.com/rescript-lang/rescript/pull/7888
@@ -42,7 +44,6 @@
 #### :rocket: New Feature
 
 - Add `Array.filterMapWithIndex` to Stdlib. https://github.com/rescript-lang/rescript/pull/7876
-- Add `littleEndian` feature for `DataView` to Stdlib. https://github.com/rescript-lang/rescript/pull/7881
 
 #### :bug: Bug fix